Add the models.py, and create the first entity, User.
The models.py file will contain all the entities in our project.
Add the User model with the attributes defined in the ER model.
Add authentication functions to the User model.
Add json support to the user model.
Add the AnonymousUser class this class will only be used if we change
from a token based auth to a username and password auth with session.