Add the cli and the shell and test commands, Add only_json function
Add the command line interface.
Add the shell command so that we can test the application within
a shell with the application context.
Add the test command so we can use and see the results of the tests
Add the only_json function that will be executed before every request
so the developers don't need to test if the request in the views that
have POST and PUT as methods have the application/json value in the
header.