Add the entity Role
The Role entity will help us add a Role Based Access Control system to the
application.
The static method insert_role is necessary because for a user to register
within our application a role will be assigned to him from the beginning and
to have this we have to have a least one role in our database.