Update User model, Update User model test case
Add the generate_confirmation_token function to the User class, this
function generates confirmation tokens with an expiration time, this
token will be sent to the user's email to verify if the email address
owner requested the account creation.
Add confirm function that verifies the token generated using the
generate_confirmation_token function and change the value of the
confirmed attribute to True if the token is valid.
Add tests to the User model test case, this tests verify the two
funcions above.