Fix authentication test, Update users test case
The api.get_user (/users/<int:id> GET) route is not longer a login
required decorated route, so we needed to change the requested route
from the test.
Remove a test from the users route test case because the route now
provides information to any user (logged in or not).